Create healthy_programmer.py
This is a mini python project which is developed as part of learning python course. This is a simple notification system based program which only runs in between office hours(i.e. in between 9:00 am to 5:00 pm). Some of the features of this project:-

it will remind user to drink water in every 20 minutes.
will remind user to exercise his/her body in every 30 minutes.
will remind user for eyes exercise in every 45 minutes.
this project contains three mp3 files for sounds of notifications and one program file.